Added contractor id to invoice details table.
Filtering on note status in invoice details view.

Updated NOT EXISTS clause in invoice details view to handle
scenario where more than 1 contractor (ie. therapist AND supervisor)
is billing for the same session.

Updated message in sp_add_new_contractors and adjusted INSERT
to include invoice frequency

Created proc for generating invoices. This proc is still a work
in progress. Commiting it and pushing, to back it up.